split into components, cleanup
