+++ web/scenes/Portal/Teams/TeamId/Apps/AppId/Configuration/AppStoreRefactored/components/FormSections/LogoSection.tsx
@@ -1,0 +1,41 @@
+import { useMemo } from "react";
+import { LogoImageUpload } from "../../LogoImageUpload";
+import { AppMetadata } from "../../types/AppStoreFormTypes";
+import { FormSection } from "../FormFields/FormSection";
+
+type LogoSectionProps = {
+  appId: string;
+  teamId: string;
+  appMetadata: AppMetadata;
+};
+
+export const LogoSection = ({
+  appId,
+  teamId,
+  appMetadata,
+}: LogoSectionProps) => {
+  const { hash } = window.location;
+  const fragmentIdentifier = useMemo(() => {
+    if (typeof window !== "undefined") {
+      return hash.substring(1);
+    }
+    return null;
+  }, [hash]);
+
+  return (
+    <FormSection
+      title="Logo image *"
+      description="Specify the logo image for your app. Image has to have a 1:1 aspect ratio."
+    >
+      <LogoImageUpload
+        appId={appId}
+        appMetadataId={appMetadata.id}
+        teamId={teamId}
+        editable={true}
+        isError={false}
+        logoFile={appMetadata.logo_img_url}
+        defaultOpen={fragmentIdentifier === "edit-logo"}
+      />
+    </FormSection>
+  );
+};
