+++ web/scenes/Portal/Teams/TeamId/Apps/AppId/Configuration/AppStoreRefactored/hooks/useSupportType.ts
@@ -1,0 +1,24 @@
+import { useCallback } from "react";
+import { UseFormSetValue } from "react-hook-form";
+import { AppStoreFormValues } from "../form-schema";
+import { SupportType } from "../types/AppStoreFormTypes";
+
+export const useSupportType = (
+  setValue: UseFormSetValue<AppStoreFormValues>,
+) => {
+  const handleSupportTypeChange = useCallback(
+    (newType: SupportType) => {
+      setValue("support_type", newType);
+      if (newType === "email") {
+        setValue("support_link", "");
+      } else {
+        setValue("support_email", "");
+      }
+    },
+    [setValue],
+  );
+
+  return {
+    handleSupportTypeChange,
+  };
+};
